401 error 해결중

+++ src/Utility/TokenRefresher.tsx
@@ -1,13 +1,12 @@
 import axios from 'axios';
 import { getCookie, removeCookie, setCookie } from './Cookie';
 
-    
-
     const refreshAPI = axios.create({
-      baseURL: "http://172.16.1.102:8080",
+      baseURL: "http://172.16.1.138:8080/reissue",
       headers: {
         "Content-Type": "application/json",
-        "Access": `${localStorage.getItem('access')}`
+        "Access" : `${localStorage.getItem('access')}`,
+        "Refresh": `${getCookie('refresh')}`
       } 
     });
 
@@ -20,6 +19,7 @@ import { getCookie, removeCookie, setCookie } from './Cookie';
       async error => {
         const originalConfig = error.config; // 기존에 수행하려고 했던 작업
         const response = error.response; // 에러 응답
+        console.log(originalConfig);
         if (response) { // 응답이 존재하는지 확인
           const msg = response.data.code; // error msg from backend
           const status = response.status; // 현재 발생한 에러 코드
@@ -30,40 +30,46 @@ import { getCookie, removeCookie, setCookie } from './Cookie';
                console.log("토큰 재발급 요청");
               try {
                 const res = await axios.post(
-                  '/reissue',{},
+                  'http://172.16.1.138:8080/reissue',{},
                   {headers: {
                     Authorization: `${localStorage.getItem('access')}`,
                     Refresh: `${getCookie('refresh')}`,
-                  }},
+                  },
+                    withCredentials: true
+                  },                  
                 );
                 // 새 토큰 저장
-                localStorage.setItem("access", res.headers.authorization);
-                setCookie("refresh", res.headers.refresh);
-    
+                localStorage.setItem("access", res.data.access);
+                setCookie('refresh', res.data.refresh);
+
                 // 새로 응답받은 데이터로 토큰 만료로 실패한 요청에 대한 인증 시도 (header에 토큰 담아 보낼 때 사용)
-                originalConfig.headers["authorization"]="Bearer "+res.headers.authorization;
-                originalConfig.headers["refresh"]= res.headers.refresh;
+                originalConfig.headers["Access"]=res.data.access;
+                originalConfig.headers["Refresh"]= res.data.refresh;
     
                 console.log("New access token obtained.");
+
+                console.log(originalConfig.headers["Access"]);
                 // 새로운 토큰으로 재요청
-                return refreshAPI(originalConfig);
+                //return refreshAPI(originalConfig);
               } catch (refreshError) {
                 console.error('An error occurred while refreshing the token:', refreshError);
               }
-            } else {
+            } else if(msg == 2) {
               localStorage.clear();
               removeCookie('refresh');
               window.alert("토큰이 만료되어 자동으로 로그아웃 되었습니다.");
             }
           } else if(status == 400 || status == 404 || status == 409) {
-            window.alert(msg); 
+            localStorage.clear();
+            removeCookie('refresh');
+            window.alert("토큰이 만료되어 자동으로 로그아웃 되었습니다.");
             console.log(msg);
           }
         } else {
           console.error('No response received:', error);
         }
-        // 다른 모든 오류를 거부하고 처리
-        return Promise.reject(error);
+        //   다른 모든 오류를 거부하고 처리
+        //return Promise.reject(error);
       },
     );
